+++ internal/ssh/ssh_tunnel.go
@@ -200,16 +200,6 @@ func (s *SSHTunnelProcessStop) DryRun(w io.Writer) error {
 	return nil
 }
 
-func resolveHost(raw string) string {
-	if raw == "" || isExplicitHost(raw) {
-		_, host, _ := SplitUserHostPort(raw)
-		return host
-	}
-
-	config := NewConfig(raw)
-	return config.host
-}
-
 func isExplicitHost(raw string) bool {
 	if strings.HasPrefix(raw, "ssh://") {
 		return true
